#4 Criando a tela da proprietario

+++ src/main/java/dev/hayann/model/Proprietario.java
@@ -18,13 +18,13 @@ public class Proprietario implements Entity {
 
     private String name;
 
-    private Integer telefone1;
+    private Long telefone1;
 
-    private Integer telefone2;
+    private Long telefone2;
 
-    private Integer telefone3;
+    private Long telefone3;
 
-    public Proprietario(Integer id, String name, Integer telefone1, Integer telefone2, Integer telefone3) {
+    public Proprietario(Integer id, String name, Long telefone1, Long telefone2, Long telefone3) {
         this.id = id;
         this.name = name;
         this.telefone1 = telefone1;
@@ -32,7 +32,7 @@ public class Proprietario implements Entity {
         this.telefone3 = telefone3;
     }
 
-    public Proprietario(String name, Integer telefone1, Integer telefone2, Integer telefone3) {
+    public Proprietario(String name, Long telefone1, Long telefone2, Long telefone3) {
         this.name = name;
         this.telefone1 = telefone1;
         this.telefone2 = telefone2;
@@ -55,27 +55,27 @@ public class Proprietario implements Entity {
         this.name = name;
     }
 
-    public Integer getTelefone1() {
+    public Long getTelefone1() {
         return telefone1;
     }
 
-    public void setTelefone1(Integer telefone1) {
+    public void setTelefone1(Long telefone1) {
         this.telefone1 = telefone1;
     }
 
-    public Integer getTelefone2() {
+    public Long getTelefone2() {
         return telefone2;
     }
 
-    public void setTelefone2(Integer telefone2) {
+    public void setTelefone2(Long telefone2) {
         this.telefone2 = telefone2;
     }
 
-    public Integer getTelefone3() {
+    public Long getTelefone3() {
         return telefone3;
     }
 
-    public void setTelefone3(Integer telefone3) {
+    public void setTelefone3(Long telefone3) {
         this.telefone3 = telefone3;
     }
